subindo

+++ NavitaAPI/Models/DTOs/PatrimonioCreateDto.cs
@@ -12,7 +12,7 @@ namespace NavitaAPI.Models.DTOs
     {
         [Required]
         public string Nome { get; set; }
-        public int NumeroTombo { get; set; }
+        public double NumeroTombo { get; set; }
         public string Descricao { get; set; }
 
         [Required]
