+++ NavitaAPI/Models/DTOs/PatrimonioUpdateDto.cs
@@ -13,7 +13,7 @@ namespace NavitaAPI.Models.DTOs
         public int Id { get; set; }
         [Required]
         public string Nome { get; set; }
-        public int NumeroTombo { get; set; }
+        public double NumeroTombo { get; set; }
         public string Descricao { get; set; }
 
         [Required]
